feat: output console log file
