+++ driving_log_replayer_cli/simulation/run.py
@@ -1,11 +1,26 @@
 from os.path import expandvars
+from pathlib import Path
 import subprocess
+import sys
 
 import termcolor
 
 from driving_log_replayer_cli.simulation.generate import TestScriptGenerator
 from driving_log_replayer_cli.simulation.result import convert
 from driving_log_replayer_cli.simulation.result import display
+
+
+def run_with_log(cmd: list, log_path: Path) -> None:
+    proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    f = log_path.open("w", encoding="utf-8")
+
+    while True:
+        line = proc.stdout.readline().decode("utf-8")
+        sys.stdout.write(line)
+        f.write(line)
+        if not line and proc.poll() is not None:
+            break
+    f.close()
 
 
 class DrivingLogReplayerTestRunner:
@@ -40,7 +55,7 @@ class DrivingLogReplayerTestRunner:
             print("aborted.")  # noqa
             return
         cmd = ["/bin/bash", generator.script_path.as_posix()]
-        subprocess.run(cmd, check=False)
+        run_with_log(cmd, Path(self.__output_directory, "console.log"))
         if output_json:
             convert(self.__output_directory)
         termcolor.cprint("<< show test result >>", "green")
